requirements.txt
allennlp==0.8.4
anytree==2.6.0
cython==0.29.1
evalai==1.3.0
h5py==2.8.0
mypy_extensions==0.4.1
nltk==3.4.3
numpy==1.15.4
pillow==6.2.0
tb-nightly
tensorboardX==1.7   
torch==1.1.0
tqdm==4.28.1
yacs==0.1.6
cffi
opencv-python
scipy
msgpack
easydict
matplotlib
pyyaml´
pycocoevalcap